Crate axvcpu

This crate provides a simple virtual CPU abstraction for hypervisors.

Structs

AxPerCpu

Host per-CPU states to run the guest.

AxVCpu

A virtual CPU with architecture-independent interface.

AxVCpuInnerMut

The mutable part of AxVCpu.

Enums

AccessWidth

The width of an access.

AxVCpuExitReason

The result of AxArchVCpu::run. Can we reference or directly reuse content from kvm-ioctls ?

VCpuState

The state of a virtual CPU.

Traits

AxArchPerCpu

Trait representing the per-CPU architecture-specific virtualization state in a virtual machine.

AxArchVCpu

A trait for architecture-specific vcpu.

Functions

clear_current_vcpu⚠

Clear the current vcpu on the current physical CPU.

get_current_vcpu

Get the current vcpu on the current physical CPU.

get_current_vcpu_mut

Get a mutable reference to the current vcpu on the current physical CPU.

set_current_vcpu⚠

Set the current vcpu on the current physical CPU.