axvcpu
Enum AxVCpuExitReason

#[non_exhaustive]
pub enum AxVCpuExitReason {

    Hypercall {
        nr: u64,
        args: [u64; 6],
    },
    MmioRead {
        addr: GuestPhysAddr,
        width: AccessWidth,
        reg: usize,
        reg_width: AccessWidth,
    },
    MmioWrite {
        addr: GuestPhysAddr,
        width: AccessWidth,
        data: u64,
    },
    IoRead {
        port: u16,
        width: AccessWidth,
    },
    IoWrite {
        port: u16,
        width: AccessWidth,
        data: u64,
    },
    ExternalInterrupt {
        vector: u64,
    },
    NestedPageFault {
        addr: GuestPhysAddr,
        access_flags: MappingFlags,
    },
    Halt,
    CpuUp {
        target_cpu: u64,
        entry_point: GuestPhysAddr,
        arg: u64,
        opaque: u64,
    },
    CpuDown {
        _state: u64,
    },
    SystemDown,
    Nothing,
    FailEntry {
        hardware_entry_failure_reason: u64,
    },

}

The result of AxArchVCpu::run. Can we reference or directly reuse content from kvm-ioctls ?

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Hypercall

The instruction executed by the vcpu performs a hypercall.

Fields

nr: u64

The hypercall number.

args: [u64; 6]

The arguments for the hypercall.

MmioRead

The instruction executed by the vcpu performs a MMIO read operation.

Fields

addr: GuestPhysAddr

The physical address of the MMIO read.

width: AccessWidth

The width of the MMIO read.

reg: usize

The index of reg to be read

reg_width: AccessWidth

The width of the reg to be read

MmioWrite

The instruction executed by the vcpu performs a MMIO write operation.

Fields

addr: GuestPhysAddr

The physical address of the MMIO write.

width: AccessWidth

The width of the MMIO write.

data: u64

The data to be written.

IoRead

The instruction executed by the vcpu performs a I/O read operation.

It’s unnecessary to specify the destination register because it’s always al, ax, or eax (as port-I/O exists only in x86).

Fields

port: u16

The port number of the I/O read.

width: AccessWidth

The width of the I/O read.

IoWrite

The instruction executed by the vcpu performs a I/O write operation.

It’s unnecessary to specify the source register because it’s always al, ax, or eax (as port-I/O exists only in x86).

Fields

port: u16

The port number of the I/O write.

width: AccessWidth

The width of the I/O write.

data: u64

The data to be written.

ExternalInterrupt

An external interrupt happened.

Note that fields may be added in the future, use .. to handle them.

Fields

vector: u64

The interrupt vector.

NestedPageFault

A nested page fault happened. (EPT violation in x86)

Note that fields may be added in the future, use .. to handle them.

Fields

addr: GuestPhysAddr

The guest physical address of the fault.

access_flags: MappingFlags

The access flags of the fault.

Halt

The vcpu is halted.

CpuUp

Try to bring up a secondary CPU.

This is used to notify the hypervisor that the target vcpu is powered on and ready to run, generally used in the boot process of a multi-core VM. This VM exit reason is architecture-specific, may be triggered by

a PSCI call in ARM

a SIPI in x86

a sbi call in RISC-V

Fields

target_cpu: u64

The target vcpu id that is to be started.

for aarch64, it contains the affinity fields of the MPIDR register,

for x86_64, it contains the APIC ID of the secondary CPU,

for RISC-V, it contains the hartid of the secondary CPU.

entry_point: GuestPhysAddr

Runtime-specified physical address of the secondary CPU’s entry point, where the vcpu can start executing.

arg: u64

This argument will be passed to the x0 register of the secondary CPU.

opaque: u64

RISC-V only, the opaque value that will be set in the a1 register when the hart starts executing at entry_point.

CpuDown

The vcpu is powered off.

This vcpu may be resumed later.

Fields

_state: u64

Currently unused. Maybe used for PSCI_POWER_STATE in the future.

SystemDown

The system should be powered off.

This is used to notify the hypervisor that the whole system should be powered off.

Nothing

Nothing special happened, the vcpu has handled the exit itself.

This exists to allow the caller to have a chance to check virtual devices/physical devices/virtual interrupts.

FailEntry

Something bad happened during VM entry, the vcpu could not be run due to unknown reasons. Further architecture-specific information is available in hardware_entry_failure_reason. Corresponds to KVM_EXIT_FAIL_ENTRY.

Fields

hardware_entry_failure_reason: u64

Architecture related VM entry failure reasons.