Trait axvcpu::AxArchVCpu

pub trait AxArchVCpu: Sized {
    type CreateConfig;
    type SetupConfig;

    // Required methods
    fn new(config: Self::CreateConfig) -> AxResult<Self>;
    fn set_entry(&mut self, entry: GuestPhysAddr) -> AxResult;
    fn set_ept_root(&mut self, ept_root: HostPhysAddr) -> AxResult;
    fn setup(&mut self, config: Self::SetupConfig) -> AxResult;
    fn run(&mut self) -> AxResult<AxVCpuExitReason>;
    fn bind(&mut self) -> AxResult;
    fn unbind(&mut self) -> AxResult;
    fn set_gpr(&mut self, reg: usize, val: usize);
}

A trait for architecture-specific vcpu.

This trait is an abstraction for virtual CPUs of different architectures.

Required Associated Types

type CreateConfig

The configuration for creating a new AxArchVCpu. Used by AxArchVCpu::new.

type SetupConfig

The configuration for setting up a created AxArchVCpu. Used by AxArchVCpu::setup.

Required Methods

fn new(config: Self::CreateConfig) -> AxResult<Self>

Create a new AxArchVCpu.

fn set_entry(&mut self, entry: GuestPhysAddr) -> AxResult

Set the entry point of the vcpu.

It’s guaranteed that this function is called only once, before AxArchVCpu::setup being called.

fn set_ept_root(&mut self, ept_root: HostPhysAddr) -> AxResult

Set the EPT root of the vcpu.

It’s guaranteed that this function is called only once, before AxArchVCpu::setup being called.

fn setup(&mut self, config: Self::SetupConfig) -> AxResult

Setup the vcpu.

It’s guaranteed that this function is called only once, after AxArchVCpu::set_entry and AxArchVCpu::set_ept_root being called.

fn run(&mut self) -> AxResult<AxVCpuExitReason>

Run the vcpu until a vm-exit occurs.

fn bind(&mut self) -> AxResult

Bind the vcpu to the current physical CPU.

fn unbind(&mut self) -> AxResult

Unbind the vcpu from the current physical CPU.

fn set_gpr(&mut self, reg: usize, val: usize)

Set the value of a general-purpose register according to the given index.

Object Safety

This trait is not object safe.