axvcpu
Struct AxVCpu

pub struct AxVCpu<A: AxArchVCpu> { /* private fields */ }

A virtual CPU with architecture-independent interface.

By delegating the architecture-specific operations to a struct implementing AxArchVCpu, this struct provides a unified interface and state management model for virtual CPUs of different architectures.

The architecture-specific operations are delegated to a struct implementing AxArchVCpu.

Note that:

This struct handles internal mutability itself, almost all the methods are &self.

This struct is not thread-safe. It’s caller’s responsibility to ensure the safety.

Implementations

impl<A: AxArchVCpu> AxVCpu<A>

pub fn new( id: usize, favor_phys_cpu: usize, phys_cpu_set: Option<usize>, arch_config: A::CreateConfig, ) -> AxResult<Self>

Create a new AxVCpu.

pub fn setup( &self, entry: GuestPhysAddr, ept_root: HostPhysAddr, arch_config: A::SetupConfig, ) -> AxResult

Setup the vcpu.

pub const fn id(&self) -> usize

Get the id of the vcpu.

pub const fn favor_phys_cpu(&self) -> usize

Get the id of the physical CPU who has the priority to run this vcpu. Currently unused.

pub const fn phys_cpu_set(&self) -> Option<usize>

Get the set of physical CPUs who can run this vcpu. If None, this vcpu has no limitation and can be scheduled on any physical CPU.

pub const fn is_bsp(&self) -> bool

Get whether the vcpu is the BSP. We always assume the first vcpu (vcpu with id #0) is the BSP.

pub fn state(&self) -> VCpuState

Get the state of the vcpu.

pub unsafe fn set_state(&self, state: VCpuState)

Set the state of the vcpu.

Safety

This method is unsafe because it may break the state transition model. Use it with caution.

pub fn with_state_transition<F, T>( &self, from: VCpuState, to: VCpuState, f: F, ) -> AxResult<T>
where F: FnOnce() -> AxResult<T>,

Execute a block with the state of the vcpu transitioned from from to to. If the current state is not from, return an error.

The state will be set to VCpuState::Invalid if an error occurs (including the case that the current state is not from).

The state will be set to to if the block is executed successfully.

pub fn with_current_cpu_set<F, T>(&self, f: F) -> T
where F: FnOnce() -> T,

Execute a block with the current vcpu set to &self.

pub fn manipulate_arch_vcpu<F, T>( &self, from: VCpuState, to: VCpuState, f: F, ) -> AxResult<T>
where F: FnOnce(&mut A) -> AxResult<T>,

Execute an operation on the architecture-specific vcpu, with the state transitioned from from to to and the current vcpu set to &self.

This method is a combination of AxVCpu::with_state_transition and AxVCpu::with_current_cpu_set.

pub fn transition_state(&self, from: VCpuState, to: VCpuState) -> AxResult

Transition the state of the vcpu. If the current state is not from, return an error.

pub fn get_arch_vcpu(&self) -> &mut A

Get the architecture-specific vcpu.

pub fn run(&self) -> AxResult<AxVCpuExitReason>

Run the vcpu.

pub fn bind(&self) -> AxResult

Bind the vcpu to the current physical CPU.

pub fn unbind(&self) -> AxResult

Unbind the vcpu from the current physical CPU.

pub fn set_gpr(&self, reg: usize, val: usize)

Sets the value of a general-purpose register according to the given index.